+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 16.0.0 Fixes
 
+- `[Card]` Added new settings and typings for card component. ([#7379](https://github.com/infor-design/enterprise/issues/7379))
 - `[General]` Added new NG 16 to the library. ([1481](https://github.com/infor-design/enterprise/issues/1481))
 
 ## 15.3.0